+++ CHANGELOG.md
@@ -183,7 +183,7 @@ More options to create `PromptbookLibrary`
 
 -   Utility `createPromptbookLibraryFromDirectory`
 -   Utility `createPromptbookLibraryFromList`
--   Add `extractOneBlockFromMarkdown` to build-in functions
+-   Add `extractBlock` to build-in functions
 -   !!! When postprocessing fails, retry in same way as failed expectations
 -   Remove problematic usage of `chalk` and use `colors` instead